pyproject.toml
[tool.poetry]
name = "media_session"
version = "0.1.0"
description = ""
authors = ["Vlad <89295404+Virashu@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.10"

winrt-windows-foundation = { version = "^2.0.0b2", platform = "win32" }
winrt-windows-foundation-collections = { version = "^2.0.0b2", platform = "win32" }
winrt-windows-media = { version = "^2.0.0b2", platform = "win32" }
winrt-windows-media-control = { version = "^2.0.0b2", platform = "win32" }
winrt-windows-storage-streams = { version = "^2.0.0b2", platform = "win32" }

dbus-python = { version = "^1.3.2", platform = "linux" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
